Remove home page
