+++ frontend/src/pages/AuthLoginPage/model/index.ts
@@ -7,94 +7,88 @@ import { routes } from "shared/routing"
 import { $session, $user } from "shared/api"
 import { chainAnonymous } from "shared/session"
 
-
 export const authLoginModel = atom(() => {
-  const currentRoute = routes.auth.login
+	const currentRoute = routes.auth.login
 
-  const loginMutation = createLoginMutation()
+	const loginMutation = createLoginMutation()
 
-  const anonymousRoute = chainAnonymous({
-    route: currentRoute,
-    otherwise: routes.home.open,
-  })
+	const anonymousRoute = chainAnonymous({
+		route: currentRoute,
+		otherwise: routes.project.list.open,
+	})
 
-  const submitted = createEvent<LoginFormSchema>()
+	const submitted = createEvent<LoginFormSchema>()
 
-  const $errorFieldEmail = createStore<string | null>(null)
-  const $errorFieldPassword = createStore<string | null>(null)
+	const $errorFieldEmail = createStore<string | null>(null)
+	const $errorFieldPassword = createStore<string | null>(null)
 
-  const $formErrors = combine({
-    email: $errorFieldEmail,
-    password: $errorFieldPassword,
-  })
+	const $formErrors = combine({
+		email: $errorFieldEmail,
+		password: $errorFieldPassword,
+	})
 
-  sample({
-    source: submitted,
-    filter: not(loginMutation.$pending),
-    target: loginMutation.start,
-  })
+	sample({
+		source: submitted,
+		filter: not(loginMutation.$pending),
+		target: loginMutation.start,
+	})
 
-  sample({
-    source: loginMutation.finished.success,
-    target: routes.home.open,
-  })
+	sample({
+		source: loginMutation.finished.success,
+		target: routes.project.list.open,
+	})
 
-  sample({
-    source: loginMutation.finished.success,
-    fn: (source) => source.result,
-    target: spread({
-      user: $user,
-      token: $session,
-    }),
-  })
+	sample({
+		source: loginMutation.finished.success,
+		fn: (source) => source.result,
+		target: spread({
+			user: $user,
+			token: $session,
+		}),
+	})
 
-  reset({
-    clock: loginMutation.finished.success,
-    target: [
-      $errorFieldEmail,
-      $errorFieldPassword,
-    ],
-  })
+	reset({
+		clock: loginMutation.finished.success,
+		target: [$errorFieldEmail, $errorFieldPassword],
+	})
 
-  sample({
-    source: loginMutation.finished.failure,
-    filter: (source) => !!source.error.data,
-    fn: (source) => {
-      if (source.error.data!.errors) {
-        const errors = source.error.data!.errors
+	sample({
+		source: loginMutation.finished.failure,
+		filter: (source) => !!source.error.data,
+		fn: (source) => {
+			if (source.error.data!.errors) {
+				const errors = source.error.data!.errors
 
-        if (errors.auth) {
-          return {
-            email: errors.auth.at(0) ?? null,
-          }
-        }
+				if (errors.auth) {
+					return {
+						email: errors.auth.at(0) ?? null,
+					}
+				}
 
-        return {
-          email: errors.email?.at(0) ?? null,
-          password: errors.password?.at(0) ?? null,
-        }
-      }
+				return {
+					email: errors.email?.at(0) ?? null,
+					password: errors.password?.at(0) ?? null,
+				}
+			}
 
-      return {
-        email: source.error.data!.message,
-        password: null,
-      }
-    },
-    target: spread({
-      email: $errorFieldEmail,
-      password: $errorFieldPassword,
-    }),
-  })
+			return {
+				email: source.error.data!.message,
+				password: null,
+			}
+		},
+		target: spread({
+			email: $errorFieldEmail,
+			password: $errorFieldPassword,
+		}),
+	})
 
-  return {
-    anonymousRoute,
+	return {
+		anonymousRoute,
 
-    loginMutation,
+		loginMutation,
 
-    submitted,
+		submitted,
 
-    $formErrors,
-  }
+		$formErrors,
+	}
 })
-
-
