+++ frontend/src/pages/AuthPasswordRecoveryPage/model/index.ts
@@ -1,10 +1,9 @@
 import { routes } from "shared/routing"
 import { chainAnonymous } from "shared/session"
 
-
 const currentRoute = routes.auth.passwordRecovery
 
 export const anonymousRoute = chainAnonymous({
-  route: currentRoute,
-  otherwise: routes.home.open,
+	route: currentRoute,
+	otherwise: routes.project.list.open,
 })
