--- frontend/src/pages/HomePage/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./HomePage"
