+++ frontend/src/pages/ProjectListPage/model/index.ts
@@ -1,9 +1,12 @@
 import { routes } from "shared/routing"
 import { chainAuthorized } from "shared/session"
+import { projectModel } from "features/current-space"
 
 const currentRoute = routes.project.list
 
 export const authorizedRoute = chainAuthorized({
-  route: currentRoute,
-  otherwise: routes.auth.login.open,
+	route: currentRoute,
+	otherwise: routes.auth.login.open,
 })
+
+export const $onboardingIsVisible = projectModel.$availableProjects.map((state) => !state.length)
